requirements.txt
customtkinter
openai-whisper
tkinterdnd2
pyinstaller
